Organise related classes into packages
