+++ src/main/java/serene/command/CommandType.java
@@ -1,4 +1,4 @@
-package serene;
+package serene.command;
 
 public enum CommandType {
     LIST, BYE, MARK, UNMARK, DELETE, TODO, DEADLINE, EVENT, EMPTY, FIND
