+++ src/main/java/serene/task/Deadline.java
@@ -1,4 +1,4 @@
-package serene;
+package serene.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
