+++ src/main/java/serene/task/TaskList.java
@@ -1,4 +1,4 @@
-package serene;
+package serene.task;
 
 import java.util.ArrayList;
