+++ src/main/java/serene/task/ToDo.java
@@ -1,4 +1,4 @@
-package serene;
+package serene.task;
 
 public class ToDo extends Task {
