+++ src/test/java/serene/EventTest.java
@@ -1,6 +1,8 @@
 package serene;
 
 import org.junit.jupiter.api.Test;
+import serene.task.Event;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class EventTest {
